Convert runner/context.js to typescript
