+++ .gitignore
@@ -8,3 +8,4 @@ runner/paths.js
 runner/plugin.js
 runner/port-scanner.js
 runner/config.js
+runner/context.js
